backend/ai_service/services/chat_api/requirements.txt
flask
flask-cors
flask-sqlalchemy
psycopg2-binary
requests
python-dotenv

langchain-core
langchain-openai
langchain-groq
langchain-google-genai
huggingface_hub